Fixed 64-bit warning.

+++ examples/common/font/text_metrics.cpp
@@ -78,7 +78,7 @@ void TextMetrics::appendText(FontHandle _fontHandle, const wchar_t* _string)
 		m_height += m_lineHeight;
 	}
 
-	for (uint32_t ii = 0, end = wcslen(_string); ii < end; ++ii)
+	for (uint32_t ii = 0, end = (uint32_t)wcslen(_string); ii < end; ++ii)
 	{
 		uint32_t codepoint = _string[ii];
 		if (m_fontManager->getGlyphInfo(_fontHandle, codepoint, glyph) )
